vendor: Re-strategize migration margins
Let's not trigger load balancing too frequently and fit more tasks in a Silver core. WALT uses the upmigrate margin as the load balancing indicator as well (for some reason) so we really want to set this to a high, yet appropriate value.

We shouldn't worry about foreground and top-app tasks since those are boosted on touch and everytime, respectively. They're going to be on Gold CPUs often without even checking for margins.

Signed-off-by: John Vincent <git@tenseventyseven.cf>