vendor: Allow RTGs to upmigrate earlier and downmigrate later
RTGs are only enabled for 'top-app' tasks and we want these to be on Gold CPUs most of the time to improve performance. Have them migrate up if a group has a demand of at least 18ms (90% of 20ms), and only downmigrate these groups if they only demand 15ms (75% of 20ms).

Signed-off-by: John Vincent <git@tenseventyseven.cf>