vendor: Experiment nuking task migration on sched boost
Triggering a massive task migration via sched_boost during latency-critical operations such as app launches is a very bad idea, and leads to noticeably-horrible stuttering[1]. But we don't exactly want to completely remove sched_boost (and not that we can remove it to begin with anyway) since we still find the slight util boost for the governor, as well as 'waking up' all Gold CPUs, still be useful on certain boost hints.

A 'sched_boost' value of 3 and a core control 'min_cpus' value of 4 on the Gold cluster is equivalent to 'sched_boost' 1 without the massive task migration boost. Do this on boost hints that trigger 'sched_boost' 1, and use just 'sched_boost' 3 for other boost hints that use just '2'.

[1]: https://github.com/Ultra119/kernel_lenovo_sdm710/commit/90a9856a4f280b3cb5043c5c374d25febbe7ba18

Signed-off-by: John Vincent <git@tenseventyseven.cf>