vendor: Keep top-app threads in the Gold cluster as much as possible
Using `top`, it seems that not all top-app threads (even the main thread of the app) still get migrated back to the Silver cluster despite UClamp minimums being set for the entire task group. This can get troublesome as the migrations don't happen opportunistically, and would induce stutters when it matters.

Since only the top-app thread group has colocation enabled, we'll keep top-app threads on the Gold cluster as much as possible by abusing RTG by setting the migration thresholds as low as possible for the RTG and early migration (which is only used by tasks in RTGs) to keep top-app threads on the Gold cluster as much as possible.

Ensure we're not thrashing battery life on standby by restoring group migrate thresholds back to stock on screen off.

Signed-off-by: John Vincent <git@tenseventyseven.cf>